scripts/requirements.txt
streamlit>=1.28.0
firebase-admin>=6.2.0
Pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0